chore: split substrapp utils
