+++ backend/api/tests/views/test_views_algo.py
@@ -14,11 +14,11 @@ from rest_framework.test import APITestCase
 
 from api.models import Algo
 from api.tests import asset_factory as factory
+from api.tests.common import AuthenticatedClient
+from api.tests.common import internal_server_error_on_exception
 from orchestrator.client import OrchestratorClient
 from orchestrator.error import OrcError
 from orchestrator.error import StatusCode
-from substrapp.tests.common import AuthenticatedClient
-from substrapp.tests.common import internal_server_error_on_exception
 from substrapp.utils import compute_hash
 
 MEDIA_ROOT = tempfile.mkdtemp()
