+++ backend/api/tests/views/test_views_compute_plan_graph.py
@@ -9,8 +9,8 @@ from rest_framework.test import APITestCase
 
 from api.models import ComputeTask
 from api.tests import asset_factory as factory
+from api.tests.common import AuthenticatedClient
 from api.views.compute_plan_graph import MAX_TASKS_DISPLAYED
-from substrapp.tests.common import AuthenticatedClient
 
 MEDIA_ROOT = tempfile.mkdtemp()
