+++ backend/api/tests/views/test_views_performance.py
@@ -13,7 +13,7 @@ from api.models import ComputePlan
 from api.models import ComputeTask
 from api.models import Performance
 from api.tests import asset_factory as factory
-from substrapp.tests.common import AuthenticatedClient
+from api.tests.common import AuthenticatedClient
 
 MEDIA_ROOT = tempfile.mkdtemp()
