+++ backend/api/tests/views/test_views_task_profiling.py
@@ -6,8 +6,8 @@ from rest_framework.test import APITestCase
 
 from api.models import ComputeTask
 from api.tests import asset_factory as factory
-from substrapp.tests.common import AuthenticatedBackendClient
-from substrapp.tests.common import AuthenticatedClient
+from api.tests.common import AuthenticatedBackendClient
+from api.tests.common import AuthenticatedClient
 
 CHANNEL = "mychannel"
 TEST_ORG = "MyTestOrg"
