+++ backend/api/views/exception_handler.py
@@ -3,8 +3,8 @@ from rest_framework import response
 from rest_framework import views
 
 import orchestrator.error
+from api.errors import ApiError
 from api.views import utils
-from substrapp import exceptions
 
 logger = structlog.get_logger(__name__)
 
@@ -17,7 +17,7 @@ def api_exception_handler(exc, context):
     if isinstance(exc, utils.ValidationExceptionError):
         return response.Response({"message": exc.data, "key": exc.key}, status=exc.st)
 
-    if isinstance(exc, exceptions.ApiError):
+    if isinstance(exc, ApiError):
         # emits a warning for all requests returning an API error response
         logger.warning("exception", class_name=exc.__class__.__name__, exception=exc)
         return exc.response()
