+++ backend/api/views/utils.py
@@ -15,11 +15,11 @@ from rest_framework.response import Response
 from rest_framework.settings import api_settings
 from rest_framework.viewsets import ViewSet
 
+from api.errors import AssetPermissionError
+from api.errors import BadRequestError
 from api.models import ComputeTask
 from organization.authentication import OrganizationUser
 from substrapp.clients import organization as organization_client
-from substrapp.exceptions import AssetPermissionError
-from substrapp.exceptions import BadRequestError
 from substrapp.storages.minio import MinioStorage
 from substrapp.utils import get_owner
